refactor: Update navigation options for room/[id] screen
