+++ app/(pages)/room/[id].tsx
